% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-bowtie.R
\name{bowtie2}
\alias{bowtie2}
\title{bowtie}
\usage{
bowtie2(opts = "", threads = 4)
}
\arguments{
\item{opts}{Options/arguments passed for bowtie index command.}

\item{threads}{Threading process to be allocated, default is 4}

\item{refgen}{string file path to Reference genome}

\item{outfile}{string for output file name}
}
\value{
Aligned sam files
}
\description{
An ultrafast memory-efficient short read aligner
}
\details{
Use combined function c() to pass in arguments as options

please refer 'bowtie2 --help' for documentation/
}
\author{
Dilip G
}
