% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-bowtie.R
\name{bowtieIndex}
\alias{bowtieIndex}
\title{bowtieIndexing}
\usage{
bowtieIndex(opts = "", threads = 4, refgen, outfile)
}
\arguments{
\item{opts}{Options/arguments passed for bowtie index command.}

\item{threads}{Threading process to be allocated, default is 4}

\item{refgen}{string file path to Reference genome}

\item{outfile}{string for output file name}
}
\value{
Indexed files and the output results for further aligning step.
}
\description{
bowtie indexing from given reference genome
}
\details{
Use combined function to make sure of bowtie arguments as options

please refer 'bowtie2-build --help' for documentation.
}
\author{
Dilip G
}
