% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-bwa.R
\name{bwaAln}
\alias{bwaAln}
\title{bwa aln}
\usage{
bwaAln(prefix, inpfile, opts)
}
\arguments{
\item{prefix}{string of prefix name to be passed for output files}

\item{inpfile}{string of input raw fastq file to be aligned}

\item{opts}{Options used for \verb{bwa aln} command.}
}
\value{
aligned sequences
}
\description{
gapped and ungapped alignment using bwa
}
\details{
perform gapped or ungapped alignment of sequences

please refer 'bwa aln' command for help and arguments options.
}
\author{
Dilip G
}
