% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-bwa.R
\name{bwaFastmap}
\alias{bwaFastmap}
\title{bwa fastmap}
\usage{
bwaFastmap(indfile, inpfile, opts)
}
\arguments{
\item{indfile}{String of index file name with path}

\item{inpfile}{string of input raw fastq file to be aligned}

\item{opts}{Options used for \verb{bwa aln} command.}
}
\value{
Quick aligned results in sam file format
}
\description{
Identify super-maximal exact matches using bwa
}
\details{
For quick and efficient aligning of sequences using SMEM

please refer 'bwa fastmap' for help and documentation.
}
\author{
Dilip G
}
