% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-bwa.R
\name{bwaIndex}
\alias{bwaIndex}
\title{bwaIndex}
\usage{
bwaIndex(file, opts = c("-a 'bwtsw'", "-p 'indexed-'"))
}
\arguments{
\item{file}{input reference fasta file to index
If the file is in gz, function will try to extract it using gunzip}

\item{opts}{Options used for \verb{bwa index} command.}
}
\value{
Stdout from bwa command and generates index files from bwa to be used for aligning
}
\description{
bwa indexing from given input fasta sequence file.
The function executes \verb{bwa index} with given opts and file as argument
}
\details{
Use combine function to write required options

-a: STR    BWT construction algorithm: bwtsw, is or rb2 \link{auto}

-p: STR    prefix of the index \link{same as fasta name}

-b: INT    block size for the bwtsw algorithm (effective with -a bwtsw) \link{10000000}

-6:        index files named as <in.fasta>.64.* instead of <in.fasta>.*

please do check bwa readme
}
\examples{
bwaIndex("human-refgen.fa", c("-p 'indexed-'"))
bwaIndex("danio-reg.fna.gz") # Indexing without any options passed
}
\author{
Dilip G
}
