% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-bwa.R
\name{bwaMem}
\alias{bwaMem}
\title{bwa}
\usage{
bwaMem(indfile, inpfile, outfile, threads = 4, opts = c(""))
}
\arguments{
\item{indfile}{string of indexed file name with path}

\item{inpfile}{string of input raw fastq file to be aligned}

\item{outfile}{string for output file name}

\item{threads}{Threading process to be allocated, default is 4}

\item{opts}{Options used for \verb{bwa mem} command.}

\item{file}{input reference fasta file to index
If the file is in gz, function will try to extract it using gunzip}
}
\value{
The aligned sequence files in sam format
}
\description{
bwa-mem to align the sequence to reference
}
\details{
The preffered algorithm to align short read sequences onto reference genome
using burrows wheeler transform with maximum exact match method.

please refer 'bwa mem' for options and arguments to be passed
}
\author{
Dilip G
}
