% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-bwa.R
\name{bwaPemerge}
\alias{bwaPemerge}
\title{bwa pemerge}
\usage{
bwaPemerge(inpfile1, inpfile2, opts)
}
\arguments{
\item{inpfile1}{string of first input fastq file name}

\item{inpfile2}{string of second input fastq file name}

\item{opts}{Options used for \verb{bwa aln} command.}
}
\value{
output results of overlapped data
}
\description{
merge overlapping paired ends
}
\details{
An experimental algorithm to find overlapping region.

please refer 'bwa pemerge' for help and documenation.
}
\author{
Dilip G
}
