% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseWrapper.R
\name{.checkBin}
\alias{.checkBin}
\title{binary command checker}
\usage{
.checkBin(bin)
}
\arguments{
\item{bin}{string indicating the executable command}
}
\value{
Boolean to check whether an executable command is available or not
}
\description{
A generic function to check for executable command in the environment.
Used to check whether tool or command exists, and stop with error message.
}
\details{
In unix system, you can refer 'man which'
}
\author{
Dilip G
}
