% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseWrapper.R
\name{.checkFiles}
\alias{.checkFiles}
\title{check File}
\usage{
.checkFiles(file)
}
\arguments{
\item{file}{path string for the files}
}
\value{
Boolean for existense of the files
}
\description{
Function to check if files exists, so command can proceed
}
\author{
Dilip G
}
