% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseWrapper.R
\name{.execBin}
\alias{.execBin}
\title{execute binary}
\usage{
.execBin(bin, args)
}
\arguments{
\item{bin}{string indicating the executable command}

\item{args}{arguments passed to the command}
}
\value{
error if command is not found, else output log of the command run
}
\description{
Function that executes the given statement in system shell environment,
if command is found. Returns the output of the commands log.
}
\author{
Dilip G
}
