% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseWrapper.R
\name{.extGz}
\alias{.extGz}
\title{GzExtractor}
\usage{
.extGz(file)
}
\arguments{
\item{file}{path string for the .gz files}
}
\value{
Extracts the gunzip file
}
\description{
Function to extract gz files using gunzip command
Not meant to be used manually, only for running indexing command we make us of this.
}
\examples{
.extGz("human-refseq.fna.gz")
}
\author{
Dilip G
}
