% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-minimap.R
\name{minimap2}
\alias{minimap2}
\title{minimap wrapper}
\usage{
minimap2(target, query, opts = "")
}
\arguments{
\item{target}{String for the indexed reference genome file name}

\item{query}{String for input fastq file for aligning.}

\item{opts}{Options or arguments to pass for minimap2 command}
}
\value{
Returns the stdout from the logs
}
\description{
mapping and alignment between DNA sequences using minimap2
}
\details{
Minimap2  is a fast sequence mapping and alignment program that can find overlaps between long noisy reads, or map long reads or their assemblies to a reference genome optionally with detailed alignment (i.e. CIGAR). At present, it works efficiently with query sequences from a few kilobases to ~100 megabases in length at a error rate ~15\%. Minimap2 outputs in the PAF or the SAM format.

Please refer the manpage for minimap2 via 'man minimap2' or 'minimap2 --help' for help and documentation
}
\author{
Dilip G
}
