% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-minimap.R
\name{minimap2Index}
\alias{minimap2Index}
\title{minimap Indexer}
\usage{
minimap2Index(save, inpfile, kmer = 15, targetram = "4G", opts = "")
}
\arguments{
\item{save}{string of filename to output the aligned results}

\item{inpfile}{input fastq file for aligning}

\item{kmer}{k-mer size, default is 15}

\item{targetram}{number of target bases to load into memory for indexing}

\item{opts}{Options or arguments to pass for minimap2 command}
}
\value{
Index file output with raw logs as displayed
}
\description{
Indexing of reference genome using minimap2
}
\details{
Indexer for aligning sequence against refernce using minimap2
}
\author{
Dilip G
}
