% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-star.R
\name{starIndex}
\alias{starIndex}
\title{STAR wrapper}
\usage{
starIndex(gendir, inpfiles, threads = 4, opts = "")
}
\arguments{
\item{gendir}{string of file path for genomic files}

\item{inpfiles}{string or combined function of input fastq files for aligning}

\item{threads}{Threading process to be allocated, default is 4}

\item{opts}{Options or arguments to be passed for 'STAR' command}
}
\value{
Aligned results of input sequences against a reference, output in sam file format.
}
\description{
RNA Spliced Transcripts Alignment to a Reference
}
\details{
An rna sequence aligner that is fast and accurate
}
\author{
Dilip G
}
